backend/functions/process_voice/requirements.txt
boto3>=1.35.0

numpy==1.26.4
faiss-cpu==1.8.0
packaging